// src/app/core/http/generated/product/v1/model/productResponse.ts
/**
 * kart-product-service API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Money } from './money';
import { RatingSummary } from './ratingSummary';
import { ProductAttributes } from './productAttributes';
import { ProductResponseCategory } from './productResponseCategory';


export interface ProductResponse { 
    sku: string;
    /**
     * Denormalized from the parent Product group (ddd-model.md\'s read-model-keying resolution).
     */
    name: string;
    description?: string;
    category: ProductResponseCategory;
    brand?: string;
    price: Money;
    status: ProductResponse.StatusEnum;
    attributes?: ProductAttributes;
    ratingSummary?: RatingSummary;
    lastUpdatedAt?: string;
}
export namespace ProductResponse {
    export type StatusEnum = 'Active' | 'Discontinued';
    export const StatusEnum = {
        Active: 'Active' as StatusEnum,
        Discontinued: 'Discontinued' as StatusEnum
    };
}